pkp/pkp-lib#2461: Do not assume the Native Import can successfully create an object in the database

+++ classes/plugins/importexport/PKPImportExportDeployment.inc.php
@@ -170,7 +170,9 @@ class PKPImportExportDeployment {
 				if (!empty($processedSubmisssionsIds)) {
 					$submissionDao = Application::getSubmissionDAO();
 					foreach ($processedSubmisssionsIds as $submissionId => $errorMessages) {
-						$submissionDao->deleteById($submissionId);
+						if ($submissionId) {
+							$submissionDao->deleteById($submissionId);
+						}
 					}
 				}
 				break;
